[REL] stock_inventory_require_lot to V11.
* [FIX] Separe models and views in diferent files.
* [FIX] Update Manifest and README
