+++ stock_inventory_require_lot/__init__.py
