+++ stock_inventory_require_lot/__manifest__.py
@@ -19,7 +19,7 @@
 ##############################################################################
 {
     'name': 'Stock Inventory Require Lot',
-    'version': '9.0.1.2.0',
+    'version': '11.0.1.0.0',
     'category': 'Warehouse Management',
     'sequence': 14,
     'summary': '',
@@ -32,11 +32,12 @@
         'stock',
     ],
     'data': [
-        'views/stock_view.xml'
+        'views/stock_change_product_views.xml',
+        'views/stock_inventory_views.xml'
     ],
     'demo': [
     ],
-    'installable': False,
+    'installable': True,
     'auto_install': False,
     'application': False,
 }
