+++ stock_inventory_require_lot/models/__init__.py
@@ -3,4 +3,5 @@
 # directory
 ##############################################################################
 
-from . import stock_inventory
+from . import stock_inventory_line
+from . import stock_change_product_qty
