+++ stock_inventory_require_lot/models/stock_change_product_qty.py
@@ -1,0 +1,10 @@
+from odoo import models, fields
+
+
+class StockChangeProductQty(models.TransientModel):
+
+    _inherit = 'stock.change.product.qty'
+
+    tracking = fields.Selection(
+        related='product_id.tracking', readonly=True
+    )
