--- stock_inventory_require_lot/models/stock_inventory.py
@@ -1,16 +1,0 @@
-from odoo import models, fields
-
-
-class StockInventoryLine(models.Model):
-
-    _inherit = 'stock.inventory.line'
-
-    tracking = fields.Selection(related='product_id.tracking', readonly=True,)
-
-
-class StockChangeProductQty(models.TransientModel):
-
-    _inherit = 'stock.change.product.qty'
-
-    tracking = fields.Selection(
-        related='product_id.tracking', readonly=True)
