+++ stock_inventory_require_lot/models/stock_inventory_line.py
@@ -1,0 +1,9 @@
+from odoo import models, fields
+
+
+class StockInventoryLine(models.Model):
+
+    _inherit = 'stock.inventory.line'
+
+    tracking = fields.Selection(
+        related='product_id.tracking', readonly=True,)
