+++ stock_inventory_require_lot/views/stock_change_product_views.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+        <record id="view_change_product_quantity" model="ir.ui.view">
+            <field name="name">Stock Change inventory</field>
+            <field name="model">stock.change.product.qty</field>
+            <field name="inherit_id" ref="stock.view_change_product_quantity"/>
+            <field name="arch" type="xml">
+                <field name="lot_id" position="after">
+                    <field name="tracking" invisible="1"/>
+                </field>
+                <field name="lot_id" position="attributes">
+                    <attribute name="attrs">{'required': [('tracking','in', ['serial', 'lot'])]}</attribute>
+                </field> 
+            </field> 
+        </record>
+</odoo>
